Linear MRCA function

Function to make the most recent common ancestor (MRCA) matrix of a tree, where entry (i,j) gives the MRCA of tips i and j. The function is linear, exploiting the fact that the tree is rooted.

linearMrca(tree, k = 0)

Arguments

tree

an object of the class phylo which should be rooted.

k

(optional) number of tips in tree, for faster computation

Examples

## generate a random tree
x <- rtree(6)

## create matrix of MRCAs: entry (i,j) is the node number of the MRCA of tips i and j
linearMrca(x,6)

#>      [,1] [,2] [,3] [,4] [,5] [,6]
#> [1,]    1    7    7    7    7    7
#> [2,]    7    2   10    9    9    8
#> [3,]    7   10    3    9    9    8
#> [4,]    7    9    9    4   11    8
#> [5,]    7    9    9   11    5    8
#> [6,]    7    8    8    8    8    6