Metric function for

Comparison of a list of trees using the Kendall Colijn metric. Output is given as a pairwise distance matrix. This is equivalent to the $D output from treespace but may be preferable for large datasets, and when principal co-ordinate analysis is not required. It includes an option to save memory at the expense of computation time.

multiDist(trees, lambda = 0, return.lambda.function = FALSE,
  save.memory = FALSE, emphasise.tips = NULL, emphasise.weight = 2)

Arguments

trees

an object of the class multiPhylo containing the trees to be compared

lambda

a number in [0,1] which specifies the extent to which topology (default, with lambda=0) or branch lengths (lambda=1) are emphasised. This argument is ignored if return.lambda.function=TRUE.

return.lambda.function

If true, a function that can be invoked with different lambda values is returned. This function returns the matrix of metric values for the given lambda.

save.memory

A flag that saves a lot of memory but increases the execution time (not compatible with return.lambda.function=TRUE).

emphasise.tips

an optional list of tips whose entries in the tree vectors should be emphasised. Defaults to NULL.

emphasise.weight

applicable only if a list is supplied to emphasise.tips, this value (default 2) is the number by which vector entries corresponding to those tips are emphasised.

Value

The pairwise tree distance matrix or a function that produces the distance matrix given a value for lambda.

Examples

## generate 10 random trees, each with 6 tips
trees <- rmtree(10,6)

## pairwise distance matrix when lambda=0
multiDist(trees)

#>        1     2     3     4     5     6     7     8     9
#> 2  7.141                                                
#> 3  3.162 7.141                                          
#> 4  4.583 6.481 5.000                                    
#> 5  4.796 6.481 5.000 3.742                              
#> 6  5.196 5.657 5.196 4.243 4.243                        
#> 7  5.196 6.325 4.796 4.243 4.000 2.828                  
#> 8  4.000 6.557 3.742 3.606 4.359 3.606 4.359            
#> 9  5.477 5.568 5.831 5.196 4.359 4.123 4.123 4.690      
#> 10 5.831 2.646 5.831 5.196 5.000 3.606 4.359 4.899 4.472

## pairwise distance matrix as a function of lambda:
m <- multiDist(trees, return.lambda.function=TRUE)

## evaluate at lambda=0. Equivalent to multiDist(trees).
m0 <- m(0)

## save memory by recomputing each tree vector for each pairwise tree comparison (for fixed lambda):
m0.5 <- multiDist(trees,0.5,save.memory=TRUE)