Web-based tree explorer

This function opens up an application in a web browser for an interactive exploration of the diversity in a set of trees. For further details please see the "help" tab within the application.

treespaceServer()

See also

Author

Thibaut Jombart thibautjombart@gmail.com

Michelle Kendall michelle.louise.kendall@gmail.com