Median transmission tree

Function to find the median of a list of transmission scenarios

wiwMedTree(matList, sampled = NULL, weights = NULL)

Arguments

matList

a list of matrices, each of which is the output of findMRCIs$mrciDepths

sampled

a vector of node IDs which corresponds to those nodes which are sampled cases

weights

optional vector of weights to correspond to the entries of matList

Value

Returns three objects:

centre: the mean of the matList entries, restricted to the sampled cases

distances: for each entry of matList, its distance from centre

mindist: the minimum of distances

median: the number of the median entry of matList, i.e. the one(s) which achieve the mindist from the centre.

Examples

# create some simple "who infected whom" scenarios:
tree1 <- cbind(Infector=1:5,Infectee=2:6)
tree2 <- cbind(Infector=c(1,5,2,2,3),Infectee=2:6)
tree3 <- cbind(Infector=c(2,2,3,4,5),Infectee=c(1,3,4,5,6))
# create list of the MRCI depth matrices:
matList <- lapply(list(tree1,tree2,tree3), function(x) findMRCIs(x)$mrciDepths)

# median tree, assuming all cases are sampled:
wiwMedTree(matList)

#> $centre
#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7]   [,8]   [,9]  [,10]  [,11]  [,12]
#> [1,] 0.3333    0    0    0    0    0    0 0.6667 0.6667 0.6667 0.6667 0.6667
#>      [,13]  [,14] [,15] [,16] [,17] [,18] [,19]  [,20] [,21] [,22] [,23] [,24]
#> [1,]     0 0.6667     2 1.333 1.667     2     0 0.6667 1.333 2.333     2     2
#>      [,25]  [,26] [,27] [,28] [,29] [,30] [,31]  [,32] [,33] [,34] [,35] [,36]
#> [1,]     0 0.6667 1.667     2     3     3     0 0.6667     2     2     3 4.333
#> 
#> $distances
#> [1] 3.180 3.432 2.963
#> 
#> $mindist
#> [1] 2.963
#> 
#> $median
#> [1] 3
#> 

# median tree when cases 1, 2 and 4 are sampled:
wiwMedTree(matList, sampled=c(1,2,4))

#> $centre
#>        [,1] [,2] [,3] [,4]   [,5]   [,6] [,7]   [,8]  [,9]
#> [1,] 0.3333    0    0    0 0.6667 0.6667    0 0.6667 2.333
#> 
#> $distances
#> [1] 0.9428 0.7454 1.3744
#> 
#> $mindist
#> [1] 0.7454
#> 
#> $median
#> [1] 2
#>